Add sf_cfg module
This module allows to manage configuration parameters. The main difference with plain DB entries is that the list of parameters is known and each parameter has a default value. External software define their parameters by adding a file to a given directory. This file contains the list of parameters and default values.